deployment/requirements.txt
streamlit
pandas
yfinance
joblib
vaderSentiment
numpy
plotly
Pillow
scikit-learn
statsmodels
transformers
